add validations to drone entity

+++ src/main/java/com/example/drones/repositories/DroneRepo.java
@@ -15,7 +15,7 @@ public interface DroneRepo extends PagingAndSortingRepository<Drone, Long> {
     Page<Drone> findAllById(Integer droneId, Pageable pageable);
 
     @Query("select d from Drone d where d.batteryCapacity >= ?1")
-    Page<Drone> findAllAvailableDrones(Double batteryCapacityMinThreshold, Pageable pageable);
+    Page<Drone> findAllAvailableDrones(Integer batteryCapacityMinThreshold, Pageable pageable);
 
     Drone findAllById(Integer droneId);
 }
